Moved structs to root level.
