+++ lib/concurrent.rb
@@ -5,17 +5,19 @@ require 'concurrent/configuration'
 require 'concurrent/atomics'
 require 'concurrent/errors'
 require 'concurrent/executors'
-require 'concurrent/struct'
 
 require 'concurrent/atomic/atomic_reference'
 require 'concurrent/async'
 require 'concurrent/dataflow'
 require 'concurrent/delay'
 require 'concurrent/future'
+require 'concurrent/immutable_struct'
 require 'concurrent/ivar'
+require 'concurrent/mutable_struct'
 require 'concurrent/mvar'
 require 'concurrent/promise'
 require 'concurrent/scheduled_task'
+require 'concurrent/settable_struct'
 require 'concurrent/synchronization_object'
 require 'concurrent/timer_task'
 require 'concurrent/tvar'
