+++ lib/concurrent/mutable_struct.rb
@@ -1,4 +1,4 @@
-require 'concurrent/struct/abstract_struct'
+require 'concurrent/collection/abstract_struct'
 require 'concurrent/synchronization_object'
 
 module Concurrent
@@ -8,7 +8,7 @@ module Concurrent
   #
   # @see http://ruby-doc.org/core-2.2.0/Struct.html Ruby standard library `Struct`
   module MutableStruct
-    include AbstractStruct
+    include Collection::AbstractStruct
 
     # @!macro [new] struct_new
     #
@@ -88,7 +88,7 @@ module Concurrent
     #   @yieldparam [Object] selfvalue the value of the member in `self`
     #   @yieldparam [Object] othervalue the value of the member in `other`
     #
-    #   @return [AbstractStruct] a new struct with the new values
+    #   @return [Collection::AbstractStruct] a new struct with the new values
     #
     #   @raise [ArgumentError] of given a member that is not defined in the struct
     def merge(other, &block)
@@ -209,7 +209,7 @@ module Concurrent
     FACTORY = Class.new(SynchronizationObject) do
       def define_struct(name, members, &block)
         synchronize do
-          clazz = AbstractStruct.define_struct_class(MutableStruct, SynchronizationObject, name, members, &block)
+          clazz = Collection::AbstractStruct.define_struct_class(MutableStruct, SynchronizationObject, name, members, &block)
           members.each_with_index do |member, index|
             clazz.send(:define_method, member) do
               synchronize { @values[index] }
