+++ spec/concurrent/mutable_struct_spec.rb
