+++ spec/concurrent/settable_struct_spec.rb
